refactor: add types.py and events.py for backend, tools, and workflow
