+++ python/beeai_framework/agents/react/runners/default/runner.py
@@ -45,8 +45,8 @@ from beeai_framework.agents.react.types import (
     ReActAgentRunIteration,
     ReActAgentTemplates,
 )
-from beeai_framework.backend.chat import ChatModelOutput
 from beeai_framework.backend.message import AssistantMessage, SystemMessage, UserMessage
+from beeai_framework.backend.types import ChatModelOutput
 from beeai_framework.emitter.emitter import EventMeta
 from beeai_framework.errors import FrameworkError
 from beeai_framework.memory.base_memory import BaseMemory
@@ -60,8 +60,9 @@ from beeai_framework.parsers.line_prefix import (
     LinePrefixParserUpdate,
 )
 from beeai_framework.retryable import Retryable, RetryableConfig, RetryableContext, RetryableInput
-from beeai_framework.tools import ToolError, ToolInputValidationError
-from beeai_framework.tools.tool import AnyTool, StringToolOutput, ToolOutput, ToolRunOptions
+from beeai_framework.tools import StringToolOutput, ToolError, ToolInputValidationError, ToolOutput
+from beeai_framework.tools.tool import AnyTool
+from beeai_framework.tools.types import ToolRunOptions
 from beeai_framework.utils.strings import create_strenum, to_json
