+++ python/beeai_framework/agents/react/types.py
@@ -18,8 +18,9 @@ from typing import Annotated, Any
 from pydantic import BaseModel, InstanceOf
 
 from beeai_framework.agents.types import AgentExecutionConfig, AgentMeta
-from beeai_framework.backend.chat import ChatModel, ChatModelOutput
+from beeai_framework.backend.chat import ChatModel
 from beeai_framework.backend.message import AnyMessage
+from beeai_framework.backend.types import ChatModelOutput
 from beeai_framework.cancellation import AbortSignal
 from beeai_framework.memory.base_memory import BaseMemory
 from beeai_framework.template import PromptTemplate
