+++ python/beeai_framework/tools/search/wikipedia.py
@@ -19,7 +19,8 @@ from pydantic import BaseModel, Field
 from beeai_framework.context import RunContext
 from beeai_framework.emitter.emitter import Emitter
 from beeai_framework.tools.search import SearchToolOutput, SearchToolResult
-from beeai_framework.tools.tool import Tool, ToolRunOptions
+from beeai_framework.tools.tool import Tool
+from beeai_framework.tools.types import ToolRunOptions
 
 
 class WikipediaToolInput(BaseModel):
