+++ python/beeai_framework/workflows/events.py
@@ -1,0 +1,41 @@
+# Copyright 2025 IBM Corp.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from typing import Generic, TypeVar
+
+from pydantic import BaseModel, InstanceOf
+
+from beeai_framework.errors import FrameworkError
+from beeai_framework.workflows.types import WorkflowRun
+
+T = TypeVar("T", bound=BaseModel)
+K = TypeVar("K", default=str)
+
+
+class WorkflowStartEvent(BaseModel, Generic[T, K]):
+    run: WorkflowRun[T, K]
+    step: K
+
+
+class WorkflowSuccessEvent(BaseModel, Generic[T, K]):
+    run: WorkflowRun[T, K]
+    state: T
+    step: K
+    next: K
+
+
+class WorkflowErrorEvent(BaseModel, Generic[T, K]):
+    run: WorkflowRun[T, K]
+    step: K
+    error: InstanceOf[FrameworkError]
