+++ python/examples/tools/custom/openlibrary.py
@@ -9,7 +9,8 @@ from beeai_framework.context import RunContext
 from beeai_framework.emitter.emitter import Emitter
 from beeai_framework.errors import FrameworkError
 from beeai_framework.tools import ToolInputValidationError
-from beeai_framework.tools.tool import JSONToolOutput, Tool, ToolRunOptions
+from beeai_framework.tools.tool import Tool
+from beeai_framework.tools.types import JSONToolOutput, ToolRunOptions
 
 
 class OpenLibraryToolInput(BaseModel):
