+++ python/examples/tools/decorator.py
@@ -12,7 +12,7 @@ from beeai_framework.backend.chat import ChatModel
 from beeai_framework.errors import FrameworkError
 from beeai_framework.logger import Logger
 from beeai_framework.memory.unconstrained_memory import UnconstrainedMemory
-from beeai_framework.tools.tool import StringToolOutput
+from beeai_framework.tools import StringToolOutput
 
 logger = Logger(__name__)
