+++ python/examples/workflows/emitter.py
@@ -7,7 +7,8 @@ from pydantic import BaseModel
 
 from beeai_framework.emitter.emitter import Emitter, EventMeta
 from beeai_framework.errors import FrameworkError
-from beeai_framework.workflows.workflow import Workflow, WorkflowReservedStepName
+from beeai_framework.workflows import WorkflowReservedStepName
+from beeai_framework.workflows.workflow import Workflow
 
 WorkflowStep: TypeAlias = Literal["pre_process", "add_loop", "post_process"]
