+++ python/examples/workflows/nesting.py
@@ -6,7 +6,8 @@ from typing import Literal, TypeAlias
 from pydantic import BaseModel
 
 from beeai_framework.errors import FrameworkError
-from beeai_framework.workflows.workflow import Workflow, WorkflowReservedStepName
+from beeai_framework.workflows import WorkflowReservedStepName
+from beeai_framework.workflows.workflow import Workflow
 
 WorkflowStep: TypeAlias = Literal["pre_process", "add_loop", "post_process"]
